Fix bug where destination variables were compiled incorrectly.
This commit fixes a bug where I incorrectly compiled instructions that
involve destinations. All variables (temporary or otherwise) are stack
allocated (alloca'd) up-front in the entry block. When an instruction with
a destination variable is executed, I had intended for the memory location
of the destination variable to be written to. However, my previous code did
not do that at all; it instead re-mapped the variable to the *LLVM
instruction's* destination in the compiler's hashtable, leading to
completely miscompiled code.