Fix string pattern match compilation
This commit fixes the incorrect compilation of string pattern matching. I had
wrongly thought that strcmp returns -1, 0, or 1; it can really return any
number.

This fix involves defining a register operand in addition to the
stack-allocated "var" operand. The Let_strcmp instruction now stores its
result in a register, and the new Let_eqint32 and Let_gtint32 instructions
do so as well. I made this change because vars are understood in terms of
the high-level type system instead of the LLVM type system, but LLVM's
integer comparison instructions return an i1, which does not exist in the
high-level type system.